additional properties don't break deserealization
As pointed out in #2, additional properties that are not defined in
the model classes lead to exceptions thrown by the normalizer. Since,
those additional properties are no backward compatibility breaks in
an API, they could be added on the remote site without any notification
and must not break the client.